Fix Size of IReadOnlyList<Dimension> to use int instead of long

+++ SharpGrad/DimensionExtender.cs
@@ -6,8 +6,8 @@ namespace SharpGrad
 {
     public static class DimensionExtender
     {
-        public static long Size(this IEnumerable<Dimension> @this)
-            => @this.Aggregate(1L, (acc, d) => acc * d.Size);
+        public static int Size(this IEnumerable<Dimension> @this)
+            => @this.Aggregate(1, (acc, d) => acc * d.Size);
 
         public static bool IsScalar(this IEnumerable<Dimension> @this)
             => !@this.Any();
